Added interceptor without spring data dependency
